+++ example/user_orders.rb
@@ -2,5 +2,6 @@ require_relative './example'
 require 'awesome_print'
 
 Example.new do |client|
-  ap client.orders
+  ords = client.orders
+  Hlpr.show_dictionary_as_table(ords,"User Orders")
 end
